[Notification] Moving Notification classes to a new Manager namespace for better organization.
