+++ src/Valkyrja/Notification/Manager/Contract/Notification.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Notification\Contract;
+namespace Valkyrja\Notification\Manager\Contract;
 
 use Valkyrja\Notification\Data\Contract\Notify;
 use Valkyrja\Notification\Entity\Contract\NotifiableUser;
