+++ tests/Unit/Notification/Provider/ServiceProviderTest.php
@@ -16,10 +16,10 @@ namespace Valkyrja\Tests\Unit\Notification\Provider;
 use PHPUnit\Framework\MockObject\Exception;
 use Valkyrja\Broadcast\Contract\Broadcaster;
 use Valkyrja\Mail\Contract\Mailer;
-use Valkyrja\Notification\Contract\Notification as Contract;
 use Valkyrja\Notification\Factory\ContainerFactory;
 use Valkyrja\Notification\Factory\Contract\Factory;
-use Valkyrja\Notification\Notification;
+use Valkyrja\Notification\Manager\Contract\Notification as Contract;
+use Valkyrja\Notification\Manager\Notification;
 use Valkyrja\Notification\Provider\ServiceProvider;
 use Valkyrja\Sms\Messenger\Contract\Messenger;
 use Valkyrja\Tests\Unit\Container\Provider\ServiceProviderTestCase;
